requirements.txt
Flask==2.2.2
Flask_Cors==4.0.0
ipython==8.11.0
openai==1.7.2
protobuf==4.25.2
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==2.2.2